pyproject.toml
[project]
version = "1.0"
name = "mnist-jax"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jax>=0.8.0",
    "pillow>=12.0.0",
    "pyarrow>=22.0.0",
    "rich>=14.2.0",
]
